% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/browseInteractions-method.R
\name{browseInteractions}
\alias{browseInteractions}
\alias{browseInteractions,ANY-method}
\alias{browseInteractions,GInteractions-method}
\alias{browseInteractions,data.frame-method}
\title{Browse genomic interactions}
\usage{
browseInteractions(object, ...)

\S4method{browseInteractions}{ANY}(object)

\S4method{browseInteractions}{GInteractions}(object, region = NULL, plot = TRUE)

\S4method{browseInteractions}{data.frame}(object, region)
}
\arguments{
\item{object}{GInteractions: Genomic interactions.}

\item{...}{not currently used.}

\item{region}{GRanges or NULL: Plotting window.}

\item{plot}{logical: Whether to produce a plot or return the plot data.}
}
\value{
ggplot or data.frame.
}
\description{
Genome browser-type plot of genomic interactions, e.g. BEDPE-like data.
}
\examples{
# TBA
}
