% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/browseIntervals-methods.R
\name{browseIntervals}
\alias{browseIntervals}
\alias{browseIntervals,ANY-method}
\alias{browseIntervals,GenomicRangesList-method}
\alias{browseIntervals,GenomicRanges-method}
\alias{browseIntervals,data.frame-method}
\title{Browse genomic intervals}
\usage{
browseIntervals(object, ...)

\S4method{browseIntervals}{ANY}(object, region = NULL)

\S4method{browseIntervals}{GenomicRangesList}(object, region = NULL, plot = TRUE)

\S4method{browseIntervals}{GenomicRanges}(object, region = NULL, plot = TRUE)

\S4method{browseIntervals}{data.frame}(object, region)
}
\arguments{
\item{object}{GRangesList, GRanges or data.frame: Genomic intervals.}

\item{...}{not currently used.}

\item{region}{GRanges or NULL: Plotting window.}

\item{plot}{logical: Whether to produce a plot or return the plot data.}
}
\value{
ggplot or data.frame.
}
\description{
Genome browser-type plot of intervals, e.g. BED-like data.
}
\details{
browseIntervals can react to metadata columns: `color`, `facet`, `thick`, `thickStart`, `thickEnd`
}
\examples{
# TBA
}
