% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/browseSegments.R
\name{browseSegments}
\alias{browseSegments}
\alias{browseSegments,ANY-method}
\alias{browseSegments,GenomicRangesList-method}
\alias{browseSegments,GenomicRanges-method}
\alias{browseSegments,data.frame-method}
\title{Browse genomic segments}
\usage{
browseSegments(object, ...)

\S4method{browseSegments}{ANY}(object, region = NULL)

\S4method{browseSegments}{GenomicRangesList}(object, region = NULL, plot = TRUE)

\S4method{browseSegments}{GenomicRanges}(object, region = NULL, plot = TRUE)

\S4method{browseSegments}{data.frame}(object, region)
}
\arguments{
\item{object}{GRangesList, GRanges or data.frame: Genomic intervals.}

\item{...}{not currently used.}

\item{region}{GRanges or NULL: Plotting window.}

\item{plot}{logical: Whether to produce a plot or return the plot data.}
}
\value{
ggplot or data.frame.
}
\description{
Genome browser-type plot of genomic segments or states.
}
\details{
browseSegments can react to metadata columns: `color`, `name`, `facet`.
}
\examples{
# TBA
}
