% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/browseSignal-methods.R
\name{browseSignal}
\alias{browseSignal}
\alias{browseSignal,ANY-method}
\alias{browseSignal,GenomicRangesList-method}
\alias{browseSignal,GenomicRanges-method}
\alias{browseSignal,data.frame-method}
\title{Browse genomic signal}
\usage{
browseSignal(object, ...)

\S4method{browseSignal}{ANY}(object)

\S4method{browseSignal}{GenomicRangesList}(object, region = NULL, plot = TRUE)

\S4method{browseSignal}{GenomicRanges}(object, region = NULL, plot = TRUE)

\S4method{browseSignal}{data.frame}(object, region)
}
\arguments{
\item{object}{GenomicRangesList, GenomicRanges or data.frame: Genomic signal.}

\item{...}{not currently used.}

\item{region}{GRanges or NULL: Plotting window.}

\item{plot}{logical: Whether to produce a plot or return the plot data.}
}
\value{
ggplot or data.frame.
}
\description{
Genome browser-type plot of genoomic signal, e.g. BigWig/bedGraph-like data.
}
\details{
browseSignal needs the following columns: `score` and can react to: `pos`, `color`, `facet`.
}
\examples{
browseSignal(ATAC_signal)
browseSignal(CAGE_CTSSs)
# See readme for more examples!
}
