% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseStack.R
\name{browseStack}
\alias{browseStack}
\title{Combine multiple browser images}
\usage{
browseStack(ggplots, strip = TRUE, ...)
}
\arguments{
\item{ggplots}{list: List of ggplots.}

\item{strip}{logical}

\item{...}{additional arguments passed to wrap_plots.}
}
\value{
Vertically aligned combined plot.
}
\description{
Combine multiple browser images
}
\examples{
#TBA
}
