% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/browseTranscripts-methods.R
\name{browseTranscripts}
\alias{browseTranscripts}
\alias{browseTranscripts,ANY-method}
\alias{browseTranscripts,GRangesList-method}
\alias{browseTranscripts,data.frame-method}
\title{Browse transcript structures}
\usage{
browseTranscripts(object, ...)

\S4method{browseTranscripts}{ANY}(object)

\S4method{browseTranscripts}{GRangesList}(object, region = NULL, CDS = NULL, plot = TRUE)

\S4method{browseTranscripts}{data.frame}(object, region)
}
\arguments{
\item{object}{GRangesList: Transcripts-per-gene.}

\item{...}{not currently used.}

\item{region}{GRanges or NULL: Plotting window.}

\item{CDS}{GRangesList: CDS-per-gene.}

\item{plot}{logical: Whether to produce a plot or return the plot data.}
}
\value{
ggplot or data.frame.
}
\description{
Genome browser-type plot of transcript-structures (intron/exon/CDS), e.g. GTF-like data.
}
\examples{
# TBA
}
