% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CAGE_TCs}
\alias{CAGE_TCs}
\alias{CAGE_CTSSs}
\alias{ATAC_peaks}
\alias{ATAC_signal}
\alias{PCHiC}
\alias{GWAS}
\title{Example plotting data}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}

An object of class \code{UnstitchedGPos} of length 276.

An object of class \code{GRanges} of length 51.

An object of class \code{GRanges} of length 3765.

An object of class \code{GInteractions} of length 97.

An object of class \code{UnstitchedGPos} of length 24.
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
CAGE_TCs

CAGE_CTSSs

ATAC_peaks

ATAC_signal

PCHiC

GWAS
}
\description{
Example datasets for plotting.
}
\keyword{datasets}
