% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/flattenRegion-methods.R
\name{flattenRegion}
\alias{flattenRegion}
\alias{flattenRegion,NULL,GenomicRanges-method}
\alias{flattenRegion,NULL,GenomicRangesList-method}
\alias{flattenRegion,NULL,GInteractions-method}
\alias{flattenRegion,GRanges,ANY-method}
\alias{flattenRegion,GInteractions,ANY-method}
\title{Easy plotting region / window extraction.}
\usage{
flattenRegion(region, object)

\S4method{flattenRegion}{`NULL`,GenomicRanges}(region, object)

\S4method{flattenRegion}{`NULL`,GenomicRangesList}(region, object)

\S4method{flattenRegion}{`NULL`,GInteractions}(region, object)

\S4method{flattenRegion}{GRanges,ANY}(region, object)

\S4method{flattenRegion}{GInteractions,ANY}(region, object)
}
\arguments{
\item{region}{GRanges: region(s) to be plotted.}

\item{object}{GenomicRanges: data to be plotted.}
}
\value{
GRanges of length 1.
}
\description{
Attempts to a single or multiple genomic ranges into a single plotting window.
}
\examples{
# TBA
}
