% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGenomeBrowser.R
\docType{package}
\name{tidyGenomeBrowser}
\alias{tidyGenomeBrowser}
\title{tidyGenomeBrowser package-wide settings.}
\description{
To avoid too many arguments to plotting functions while still retaining flexibility, some plotting parameters can be changed using package-wide settings.
}
\section{Package-wide settings}{

\itemize{
\item tidyGenomeBrowser.prefix: Prefix to seqlevel on X-axis.
\item tidyGenomeBrowser.strand: Strand colors for plus, minus and unstranded.
\item tidyGenomeBrowser.interactions: Colors or interactions arches inside/outside of plotting window in browseInteractions.
\item tidyGenomeBrowser.alpha: Transparency for browseSignal and browsePositions.
\item tidyGenomeBrowser.nudge: How much to nudge names upwards relative to the feature start in browseIntervals and browseTranscripts.
\item tidyGenomeBrowser.wiggle: Extra distance around each feature when binning features along the Y-axis in browseIntervals, browsePosition and browseTranscripts.
\item tidyGenomeBrowser.height: Height of features in browseIntervals.
\item tidyGenomeBrowser.thick: Height of thick features in browseIntervals.
\item tidyGenomeBrowser.flip: Whether to flip signal on the minus strand to negative values in browseSignal.
\item tidyGenomeBrowser.name: Whether to add names to features in browseIntervals and browseTranscripts.
\item tidyGenomeBrowser.tx: Line-thickness of Intron, Exon and CDS regions in browseTranscripts.
}
}

